backend/pyproject.toml
[project]
name = "scolo-backend"
version = "0.1.0"
description = "Scolo Agent Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "modal>=0.73.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
